moving fnpr_group_plot to bias_detection
